[videre] Allow to render a char with a specified color Add a widget Character to quickly render a char Row available dimension to int in Row and Column Use Div for check buttons

+++ videre/layouts/column.py
@@ -76,7 +76,9 @@ class Column(AbstractControlsLayout):
                 for i, ctrl in to_render:
                     if total_height >= height:
                         break
-                    available_height = (remaining_height * weights[i]) // total_weight
+                    available_height = int(
+                        (remaining_height * weights[i]) // total_weight
+                    )
                     surface = ctrl.render(window, w_hint, available_height)
                     rendered[i] = (ctrl, surface)
                     sizes[i] = available_height
